Weatherella/requirements.txt
requests>=2.28
pytest>=7.0
python-dotenv>=1.0
flask>=2.0.0
flask-cors>=3.0.10